+++ .gitignore
@@ -1,6 +1,4 @@
 node_modules
-dist
-out
 .turbo
 .DS_Store
 *.local
