+++ package.json
@@ -16,7 +16,7 @@
 		"lefthook": "^1.6.16",
 		"turbo": "^2.0.4"
 	},
-	"packageManager": "npm@10.8.1+sha256.b8807aebb9656758e2872fa6e7c564b506aa2faa9297439a478d471d2fe32483",
+	"packageManager": "pnpm@9.4.0+sha256.b6fd0bfda555e7e584ad7e56b30c68b01d5a04f9ee93989f4b93ca8473c49c74",
 	"volta": {
 		"node": "22.3.0"
 	}
