--- packages/extension/tsconfig.json
@@ -1,6 +1,0 @@
-{
-	"extends": "@tsconfig/vite-react",
-	"compilerOptions": {
-		"types": ["vite/client"]
-	}
-}
