+++ packages/extension/wxt.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "wxt";
+
+export default defineConfig({
+	manifest: {
+		name: "Comment Overlay Helper",
+		permissions: ["tabs", "scripting", "storage"],
+		action: {},
+	},
+});
