+++ packages/types/package.json
@@ -1,0 +1,8 @@
+{
+	"name": "@repo/types",
+	"version": "0.0.0",
+	"private": true,
+	"exports": {
+		".": "./src/index.ts"
+	}
+}
