+++ packages/types/src/helper.d.ts
@@ -1,0 +1,13 @@
+/** コメントペイロード */
+export type HelperCommentPayload = {
+	type: "comment";
+	text: string;
+};
+
+/** リアクションペイロード */
+export type HelperReactionPayload = {
+	type: "reaction";
+	src: string;
+	left: string | undefined;
+	size: string | undefined;
+};
